integrate confirmIdenticity functions to checkUnicity_script
