+++ Framework/Form.php
@@ -34,7 +34,6 @@ class Form {
     private $readonly;
     private $checkUnicity;
     private $suggestLogin;
-    private $confirmIdenticity;
 
     private $choices;
     private $choicesid;
@@ -315,7 +314,7 @@ class Form {
      * @param string $value Input default value
      */
     // #105: add readonly
-    public function addText($name, $label, $isMandatory = false, $value = "", $enabled = "", $readonly = "", $checkUnicity = false, $suggestLogin = false, $confirmIdenticity = false) {
+    public function addText($name, $label, $isMandatory = false, $value = "", $enabled = "", $readonly = "", $checkUnicity = false, $suggestLogin = false) {
         $this->types[] = "text";
         $this->names[] = $name;
         $this->labels[] = $label;
@@ -331,9 +330,6 @@ class Form {
         $this->checkUnicity[] = $checkUnicity;
         if ($suggestLogin) {
             $this->suggestLogin = true;
-        }
-        if ($confirmIdenticity) {
-            $this->confirmIdenticity = $confirmIdenticity;
         }
     }
 
@@ -444,7 +440,7 @@ class Form {
      * @param string $isMandatory True if mandatory input
      * @param string $value Input default value
      */
-    public function addEmail($name, $label, $isMandatory = false, $value = "", $checkUnicity = false, $confirmIdenticity = false) {
+    public function addEmail($name, $label, $isMandatory = false, $value = "", $checkUnicity = false) {
         $this->types[] = "email";
         $this->names[] = $name;
         $this->labels[] = $label;
@@ -458,9 +454,6 @@ class Form {
         $this->submitOnChange[] = false;
         $this->readonly[] = false;
         $this->checkUnicity[] = $checkUnicity;
-        if ($confirmIdenticity) {
-            $this->confirmIdenticity = $confirmIdenticity;
-        }
     }
 
     /**
@@ -776,10 +769,6 @@ class Form {
         
         if ($this->suggestLogin) {
             $html .= $formHtml->suggestLoginScript();
-        }
-
-        if ($this->confirmIdenticity) {
-            $html .= $formHtml->confirmIdenticityScript();
         }
 
         if ($this->isFormAdd === true) {
